Update Upload.pm
This needs to be tested but I made it so that it will report as finished if the output says 100% not just if it finishes outputing to the log file. I guess it never stops tailing the file even if the gtupload has finishe and is then waiting for more output. This should fix that problem.